Move construct_oligo; update terminology; add comments
Move construct_oligo from guide_search/primer_search subclasses to search.py; change "do_not_memoize_guide" to "do_not_memoize_oligo"; change "ignored_kmers" to "disallowed_kmers"; change all references to guide in search.py to oligo; add comments to is_symmetric and is_complement
