+++ adapt/primer_search.py
@@ -132,9 +132,7 @@ class PrimerSearcherMinimizePrimers(search.OligoSearcherMinimizeNumber):
                     GC content
                 """
                 gc_frac = oligo.gc_frac(primer_seq)
-                if gc_frac < lo or gc_frac > hi:
-                    return False
-                return True
+                return (gc_frac >= lo and gc_frac <= hi)
             is_suitable_fns = is_suitable_fns + [check_gc_content]
 
         super().__init__(aln=aln, min_oligo_length=primer_length,
@@ -142,288 +140,6 @@ class PrimerSearcherMinimizePrimers(search.OligoSearcherMinimizeNumber):
             mismatches=mismatches, missing_data_params=missing_data_params,
             is_suitable_fns=is_suitable_fns, **kwargs)
 
-    def construct_oligo(self, start, oligo_length, seqs_to_consider,
-            num_needed=None, stop_early=True):
-        """Construct a single primer to target a set of sequences in the alignment.
-
-        This constructs a primer to target sequence within the range [start,
-        start+primer_length]. It only considers the sequences with indices given in
-        seqs_to_consider.
-
-        Args:
-            start: start position in alignment at which to target
-            seqs_to_consider: dict mapping universe group ID to collection of
-                indices to use when constructing the primer
-            num_needed: dict mapping universe group ID to the number of sequences
-                from the group that are left to cover in order to achieve
-                a desired coverage; these are used to help construct a
-                primer
-            stop_early: if True, impose early stopping criteria while iterating
-                over clusters to improve runtime
-
-        Returns:
-            tuple (x, y, z) where:
-                x is the sequence of the constructed primer
-                y is a set of indices of sequences (a subset of
-                    values in seqs_to_consider) to which the primer x will
-                    hybridize
-                z is the marginal contribution of the primer to the objective
-            (Note that it is possible that x binds to no sequences and that
-            y will be empty.)
-        """
-        # TODO: There are several optimizations that can be made to
-        # this function that take advantage of G-U pairing in order
-        # to lower the number of primers that need to be designed.
-        # Two are:
-        #  1) The function SequenceClusterer.cluster(..), which is
-        #     used here, can cluster accounting for G-U pairing (e.g.,
-        #     such that 'A' hashes to 'G' and 'C' hashes to 'T',
-        #     so that similar primer sequences hash to the same
-        #     value tolerating G-U similarity).
-        #  2) Instead of taking a consensus sequence across a
-        #     cluster, this can take a pseudo-consensus that
-        #     accounts for G-U pairing (e.g., in which 'A' is
-        #     treated as 'G' and 'C' is treated as 'T' for
-        #     the purposes of generating a consensus sequence).
-
-        assert start + oligo_length <= self.aln.seq_length
-        assert len(seqs_to_consider) > 0
-
-        for pos in range(start, start + oligo_length):
-            if self.aln.frac_missing_at_pos(pos) > self.missing_threshold:
-                raise search.CannotConstructOligoError(("Too much missing "
-                    "data at a position in the target range"))
-
-        aln_for_primer = self.aln.extract_range(start, start + oligo_length)
-
-        if self.predictor is not None:
-            # Extract the target sequences, including context to use with
-            # prediction
-            if (start - self.predictor.context_nt < 0 or
-                    start + oligo_length + self.predictor.context_nt >
-                    self.aln.seq_length):
-                raise search.CannotConstructOligoError(("The context needed "
-                    "for the target to predict activity falls outside the "
-                    "range of the alignment at this position"))
-            aln_for_primer_with_context = self.aln.extract_range(
-                    start - self.predictor.context_nt,
-                    start + oligo_length + self.predictor.context_nt)
-
-        # Before modifying seqs_to_consider, make a copy of it
-        seqs_to_consider_cp = {}
-        for group_id in seqs_to_consider.keys():
-            seqs_to_consider_cp[group_id] = set(seqs_to_consider[group_id])
-        seqs_to_consider = seqs_to_consider_cp
-
-        all_seqs_to_consider = set.union(*seqs_to_consider.values())
-
-        # Ignore any sequences in the alignment that have a gap in
-        # this region
-        if self.predictor is not None:
-            seqs_with_gap = set(aln_for_primer_with_context.seqs_with_gap(
-                all_seqs_to_consider))
-        else:
-            seqs_with_gap = set(aln_for_primer.seqs_with_gap(
-                all_seqs_to_consider))
-        for group_id in seqs_to_consider.keys():
-            seqs_to_consider[group_id].difference_update(seqs_with_gap)
-        all_seqs_to_consider = set.union(*seqs_to_consider.values())
-
-        # Only consider sequences in the alignment that have the
-        # required flanking sequence(s)
-        seqs_with_flanking = self.aln.seqs_with_required_flanking(
-            start, oligo_length, self.required_flanking_seqs,
-            seqs_to_consider=all_seqs_to_consider)
-        for group_id in seqs_to_consider.keys():
-            seqs_to_consider[group_id].intersection_update(seqs_with_flanking)
-        all_seqs_to_consider = set.union(*seqs_to_consider.values())
-
-        # If every sequence in this region has a gap or does not contain
-        # required flanking sequence, then there are none left to consider
-        if len(all_seqs_to_consider) == 0:
-            raise search.CannotConstructOligoError(("All sequences in region "
-                "have a gap and/or do not contain required flanking sequences"))
-
-        seq_rows = aln_for_primer.make_list_of_seqs(all_seqs_to_consider,
-            include_idx=True)
-        if self.predictor is not None:
-            seq_rows_with_context = aln_for_primer_with_context.make_list_of_seqs(
-                    all_seqs_to_consider, include_idx=True)
-
-        if self.predictor is not None:
-            # Memoize activity evaluations
-            pair_eval = {}
-        def determine_binding_and_active_seqs(pr_sequence):
-            binding_seqs = set()
-            num_bound = 0
-            if self.predictor is not None:
-                num_passed_predict_active = 0
-                # Determine what calls to make to
-                # self.predictor.determine_highly_active(); it
-                # is best to batch these
-                pairs_to_eval = []
-                for i, (seq, seq_idx) in enumerate(seq_rows):
-                    if oligo.binds(pr_sequence, seq, self.mismatches,
-                            self.allow_gu_pairs):
-                        seq_with_context, _ = seq_rows_with_context[i]
-                        pair = (seq_with_context, pr_sequence)
-                        pairs_to_eval += [pair]
-                # Evaluate activity
-                evals = self.predictor.determine_highly_active(start, pairs_to_eval)
-                for pair, y in zip(pairs_to_eval, evals):
-                    pair_eval[pair] = y
-                # Fill in binding_seqs
-                for i, (seq, seq_idx) in enumerate(seq_rows):
-                    if oligo.binds(pr_sequence, seq, self.mismatches,
-                            self.allow_gu_pairs):
-                        num_bound += 1
-                        seq_with_context, _ = seq_rows_with_context[i]
-                        pair = (seq_with_context, pr_sequence)
-                        if pair_eval[pair]:
-                            num_passed_predict_active += 1
-                            binding_seqs.add(seq_idx)
-            else:
-                num_passed_predict_active = None
-                for i, (seq, seq_idx) in enumerate(seq_rows):
-                    if oligo.binds(pr_sequence, seq, self.mismatches,
-                            self.allow_gu_pairs):
-                        num_bound += 1
-                        binding_seqs.add(seq_idx)
-            return binding_seqs, num_bound, num_passed_predict_active
-
-        # Define a score function (higher is better) for a collection of
-        # sequences covered by a primer
-        if num_needed is not None:
-            # This is the number of sequences it contains that are needed to
-            # achieve the partial cover; we can compute this by summing over
-            # the number of needed sequences it contains, taken across the
-            # groups in the universe
-            # Memoize the scores because this computation might be expensive
-            seq_idxs_scores = {}
-            def seq_idxs_score(seq_idxs):
-                seq_idxs = set(seq_idxs)
-                tc = tuple(seq_idxs)
-                if tc in seq_idxs_scores:
-                    return seq_idxs_scores[tc]
-                score = 0
-                for group_id, needed in num_needed.items():
-                    contained_in_seq_idxs = seq_idxs & seqs_to_consider[group_id]
-                    score += min(needed, len(contained_in_seq_idxs))
-                seq_idxs_scores[tc] = score
-                return score
-        else:
-            # Score by the number of sequences it contains
-            def seq_idxs_score(seq_idxs):
-                return len(seq_idxs)
-
-        # First construct the optimal primer to cover the sequences. This would be
-        # a string x that maximizes the number of sequences s_i such that x and
-        # s_i are equal to within 'mismatches' mismatches; it's called the "max
-        # close string" or "close to most strings" problem. For simplicity, let's
-        # do the following: cluster the sequences (just the portion with
-        # potential primers) with LSH, choose a primer for each cluster to be
-        # the consensus of that cluster, and choose the one (across clusters)
-        # that has the highest score
-        if self.clusterer is None:
-            # Don't cluster; instead, draw the consensus from all the
-            # sequences. Effectively, treat it as there being just one
-            # cluster, which consists of all sequences to consider
-            clusters_ordered = [all_seqs_to_consider]
-        else:
-            # Cluster the sequences
-            clusters = self.clusterer.cluster(seq_rows)
-
-            # Sort the clusters by score, from highest to lowest
-            # Here, the score is determined by the sequences in the cluster
-            clusters_ordered = sorted(clusters, key=seq_idxs_score, reverse=True)
-
-        best_pr = None
-        best_pr_binding_seqs = None
-        best_pr_score = 0
-        stopped_early = False
-        for cluster_idxs in clusters_ordered:
-            if stop_early and best_pr_score > seq_idxs_score(cluster_idxs):
-                # The primer from this cluster is unlikely to exceed the current
-                # score; stop early
-                stopped_early = True
-                break
-
-            pr = aln_for_primer.determine_consensus_sequence(
-                cluster_idxs)
-            if 'N' in pr:
-                # Skip this; all sequences at a position in this cluster
-                # are 'N'
-                continue
-            skip_cluster = False
-            for is_suitable_fn in self.is_suitable_fns:
-                if is_suitable_fn(pr) is False:
-                    # Skip this cluster
-                    skip_cluster = True
-                    break
-            if skip_cluster:
-                continue
-            # Determine the sequences that are bound by this primer (and
-            # where it is 'active', if self.predictor is set)
-            binding_seqs, num_bound, num_passed_predict_active = \
-                    determine_binding_and_active_seqs(pr)
-            score = seq_idxs_score(binding_seqs)
-            if score > best_pr_score:
-                best_pr = pr
-                best_pr_binding_seqs = binding_seqs
-                best_pr_score = score
-
-            # Impose an early stopping criterion if self.predictor is
-            # used, because using it is slow
-            if self.predictor is not None and stop_early:
-                if (num_bound >= 0.5*len(cluster_idxs) and
-                        num_passed_predict_active < 0.1*len(cluster_idxs)):
-                    # pr binds (according to binds()) many
-                    # sequences, but is not predicted to be active against
-                    # many; it is likely that this region is poor according to
-                    # the self.predictor (e.g., due to sequence composition).
-                    # Rather than trying other clusters at this site, just skip
-                    # it. Note that this is just a heuristic; it can help
-                    # runtime when the self.predictor is used, but is not needed
-                    # and may hurt the optimality of the solution
-                    stopped_early = True
-                    break
-        pr = best_pr
-        binding_seqs = best_pr_binding_seqs
-        score = best_pr_score
-
-        # It's possible that the consensus sequence (primer) of no cluster
-        # binds to any of the sequences. In this case, simply go through all
-        # sequences and find the first that has no ambiguity and is suitable
-        # and active, and make this the primer
-        if pr is None and not stopped_early:
-            for i, (s, idx) in enumerate(seq_rows):
-                if not set(s).issubset(set(['A', 'C', 'G', 'T'])):
-                    # s has ambiguity; skip it
-                    continue
-                skip_cluster = False
-                for is_suitable_fn in self.is_suitable_fns:
-                    if is_suitable_fn(s) is False:
-                        # Skip this cluster
-                        skip_cluster = True
-                        break
-                if skip_cluster:
-                    continue
-                if self.predictor is not None:
-                    s_with_context, _ = seq_rows_with_context[i]
-                    if not self.predictor.determine_highly_active(start,
-                            [(s_with_context, s)])[0]:
-                        # s is not active against itself; skip it
-                        continue
-                # s has no ambiguity and is a suitable primer; use it
-                pr = s
-                binding_seqs, _, _ = determine_binding_and_active_seqs(pr)
-                break
-
-        if pr is None:
-            raise search.CannotConstructOligoError(("No primers are suitable "
-                "or active"))
-
-        return (pr, binding_seqs, score)
 
     def find_primers(self, max_at_site=None):
         """Find primers across the alignment.
