+++ adapt/tests/test_guide_search.py
@@ -535,255 +535,6 @@ class TestGuideSearcherMinimizeGuides(unittest.TestCase):
         # All sequences are bound by a guide
         self.assertEqual(gs.total_frac_bound(guides), 1.0)
 
-    def test_construct_oligo_a(self):
-        self.a.clusterer = self.gc
-        self.assertEqual(self.a.construct_oligo(0, 4, {0: {0,1,2,3,4}}),
-                         ('ATCG', {0,1,2,3}, 4))
-        self.assertIn(self.a.construct_oligo(0, 4, {0: {2,3}}),
-                      [('ATCG', {2,3}, 2), ('ACCG', {2,3}, 2)])
-        self.assertEqual(self.a.construct_oligo(1, 4, {0: {0,1,2,3,4}}),
-                         ('TCGA', {0,1,2,3}, 4))
-        self.assertEqual(self.a.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAA', {0,2,4}, 3))
-        self.assertIn(self.a.construct_oligo(2, 4, {0: {0,1,2,3}}),
-                      [('CGAA', {0,2}, 2), ('CGAT', {1,3}, 2)])
-        self.a.mismatches = 1
-        self.assertEqual(self.a.construct_oligo(0, 4, {0: {0,1,2,3,4}}),
-                         ('ATCG', {0,1,2,3,4}, 5))
-        self.assertEqual(self.a.construct_oligo(0, 4, {0: {4}}),
-                         ('AGCG', {4}, 1))
-        self.assertEqual(self.a.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAA', {0,1,2,3,4}, 5))
-        self.assertIn(self.a.construct_oligo(2, 4, {0: {0,1,2,3}}),
-                      [('CGAA', {0,1,2,3}, 4), ('CGAT', {0,1,2,3}, 4)])
-        self.a.mismatches = 2
-        self.assertEqual(self.a.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAA', {0,1,2,3,4}, 5))
-
-    def test_construct_oligo_b(self):
-        self.b.clusterer = self.gc
-        with self.assertRaises(search.CannotConstructOligoError):
-            # Should fail when the only sequence given (1) has an indel
-            self.b.construct_oligo(0, 4, {0: {1}})
-
-    def test_construct_oligo_ambiguous(self):
-        # Alignment has many Ns, which makes it difficult to write test cases
-        # when clustering (the clusters tend to consist of guides in
-        # which a position only has N); so pass None to guide_clusterer in
-        # construct_oligo() to skip clustering
-        seqs = ['ATCGAA', 'ATNNAT', 'ATCGNN', 'ATNNAT', 'ATNNAC']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        self.assertEqual(self.gen_gs.construct_oligo(0, 4, {0: {0,1,2,3,4}}),
-                         ('ATCG', {0,2}, 2))
-        self.assertEqual(self.gen_gs.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAA', {0}, 1))
-        with self.assertRaises(search.CannotConstructOligoError):
-            # Should fail when 'N' is all that exists at a position
-            self.gen_gs.construct_oligo(0, 4, {0: {1,3,4}})
-        self.gen_gs.mismatches = 1
-        self.assertEqual(self.gen_gs.construct_oligo(0, 4, {0: {0,1,2,3,4}}),
-                         ('ATCG', {0,2}, 2))
-        self.assertEqual(self.gen_gs.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAT', {0}, 1))
-        with self.assertRaises(search.CannotConstructOligoError):
-            # Should fail when a potential guide (here, 'CGAC') cannot
-            # bind to any sequence because they all have 'N' somewhere
-            self.gen_gs.construct_oligo(2, 4, {0: {2,4}})
-        self.gen_gs.mismatches = 2
-        self.assertEqual(self.gen_gs.construct_oligo(0, 4, {0: {0,1,2,3,4}}),
-                         ('ATCG', {0,1,2,3,4}, 5))
-        self.assertEqual(self.gen_gs.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAT', {0,1,2,3}, 4))
-        self.assertIn(self.gen_gs.construct_oligo(2, 4, {0: {2,3,4}}),
-                      [('CGAC', {2,4}, 2), ('CGAT', {2,3}, 2)])
-        self.assertEqual(self.gen_gs.construct_oligo(2, 4, {0: {2,4}}),
-                         ('CGAC', {2,4}, 2))
-        self.gen_gs.mismatches = 3
-        self.assertEqual(self.gen_gs.construct_oligo(2, 4, {0: {0,1,2,3,4}}),
-                         ('CGAT', {0,1,2,3,4}, 5))
-
-    def test_construct_oligo_with_large_group_needed(self):
-        seqs = ['ATCGAA',
-                'ATCGAA',
-                'GGGCCC',
-                'ATCGAA',
-                'ATCGAA',
-                'ATCGAA',
-                'GGGCCC']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        self.gen_gs.clusterer = self.gc
-
-        seqs_to_consider = {0: {0, 1, 3, 4, 5}, 1: {2, 6}}
-        num_needed = {0: 3, 1: 1}
-        # 'ATCGAA' is most sequences, and let's construct a guide by
-        # needing more from the group consisting of these sequences
-        self.assertEqual(self.gen_gs.construct_oligo(0, 4, seqs_to_consider,
-                                                     num_needed=num_needed),
-                         ('ATCG', {0, 1, 3, 4, 5}, 3))
-
-    def test_construct_oligo_with_small_group_needed(self):
-        seqs = ['ATCGAA',
-                'ATCGAA',
-                'GGGCCC',
-                'ATCGAA',
-                'ATCGAA',
-                'ATCGAA',
-                'GGGCCC']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        self.gen_gs.clusterer = self.gc
-
-        seqs_to_consider = {0: {0, 1, 3, 4, 5}, 1: {2, 6}}
-        num_needed = {0: 1, 1: 2}
-        # 'ATCGAA' is most sequences, but let's construct a guide by
-        # needing more from a group consisting of the 'GGGCCC' sequences
-        self.assertEqual(self.gen_gs.construct_oligo(0, 4, seqs_to_consider,
-                                                     num_needed=num_needed),
-                         ('GGGC', {2, 6}, 2))
-
-    def test_construct_oligo_with_suitable_fn(self):
-        seqs = ['GTATCAAAT',
-                'CTACCAAAA',
-                'GTATCAAAT',
-                'GTATCAAAT']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        guide_length = 6
-        self.gen_gs.min_oligo_length = guide_length
-        self.gen_gs.max_oligo_length = guide_length
-        seqs_to_consider = {0: {0, 1, 2, 3}}
-        self.gen_gs.clusterer = alignment.SequenceClusterer(
-            lsh.HammingDistanceFamily(guide_length), k=3)
-        self.gen_gs.mismatches = 1
-
-        # The best guide is 'GTATCA'
-        self.assertEqual(self.gen_gs.construct_oligo(0, guide_length,
-                                                     seqs_to_consider),
-                         ('GTATCA', {0, 2, 3}, 3))
-
-        # Do not allow guides with 'TAT' in them
-        def f(guide):
-            if 'TAT' in guide:
-                return False
-            else:
-                return True
-        prev_suitable_fns = self.gen_gs.is_suitable_fns
-        self.gen_gs.is_suitable_fns = prev_suitable_fns + [f]
-
-        # Now the best guide is 'CTACCA'
-        self.assertEqual(self.gen_gs.construct_oligo(0, guide_length,
-                                                     seqs_to_consider),
-                         ('CTACCA', {1}, 1))
-
-        # Do not allow guides with 'A' in them
-        def f(guide):
-            if 'A' in guide:
-                return False
-            else:
-                return True
-        self.gen_gs.is_suitable_fns = prev_suitable_fns + [f]
-
-        # Now there is no suitable guide
-        with self.assertRaises(search.CannotConstructOligoError):
-            self.gen_gs.construct_oligo(0, guide_length, seqs_to_consider)
-
-    def test_construct_oligo_with_predictor(self):
-        seqs = ['GTATCAAAT',
-                'ATACCAAAA',
-                'GTATCAAAT',
-                'GTATCAAAT']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        guide_length = 6
-        self.gen_gs.min_oligo_length = guide_length
-        self.gen_gs.max_oligo_length = guide_length
-        seqs_to_consider = {0: {0, 1, 2, 3}}
-        self.gen_gs.clusterer = alignment.SequenceClusterer(
-            lsh.HammingDistanceFamily(guide_length), k=3)
-        self.gen_gs.mismatches = 1
-
-        self.assertEqual(self.gen_gs.construct_oligo(0, guide_length, seqs_to_consider),
-                         ('GTATCA', {0, 2, 3}, 3))
-
-        # Only predict guides starting with 'A' to be active
-        class PredictorTest:
-            def __init__(self):
-                self.context_nt = 0
-            def determine_highly_active(self, start_pos, pairs):
-                y = []
-                for target, guide in pairs:
-                    y += [guide[0] == 'A']
-                return y
-        self.gen_gs.predictor = PredictorTest()
-        # Now the best guide is 'ATACCA'
-        self.assertEqual(self.gen_gs.construct_oligo(0, guide_length,
-                                                     seqs_to_consider,
-                                                     stop_early=False),
-                         ('ATACCA', {1}, 1))
-
-        # Only predict guides starting with 'A' to be active, and impose an
-        # early stopping criterion
-        class PredictorTest:
-            def __init__(self):
-                self.context_nt = 0
-            def determine_highly_active(self, start_pos, pairs):
-                y = []
-                for target, guide in pairs:
-                    y += [guide[0] == 'A']
-                return y
-        self.gen_gs.predictor = PredictorTest()
-        # With early stopping, it will not find a guide
-        with self.assertRaises(search.CannotConstructOligoError):
-            self.gen_gs.construct_oligo(0, guide_length, seqs_to_consider)
-
-        # Only predictor guides starting with 'C' to be active
-        class PredictorTest:
-            def __init__(self):
-                self.context_nt = 0
-            def determine_highly_active(self, start_pos, pairs):
-                y = []
-                for target, guide in pairs:
-                    y += [guide[0] == 'C']
-                return y
-        self.gen_gs.predictor = PredictorTest()
-        # Now there is no suitable guide
-        with self.assertRaises(search.CannotConstructOligoError):
-            self.gen_gs.construct_oligo(0, guide_length, seqs_to_consider)
-
-    def test_construct_oligo_with_required_flanking(self):
-        seqs = ['TCAAAT',
-                'CCAAAA',
-                'CATTTT',
-                'CATTTT',
-                'CATTTT',
-                'CATTTT',
-                'CATTTT',
-                'CATTTT',
-                'CATTTT',
-                'TCAAAT',
-                'TCAAAT']
-        self.gen_gs.aln = alignment.Alignment.from_list_of_seqs(seqs)
-        guide_length = 2
-        self.gen_gs.min_oligo_length = guide_length
-        self.gen_gs.max_oligo_length = guide_length
-        self.gen_gs.clusterer = alignment.SequenceClusterer(
-            lsh.HammingDistanceFamily(guide_length),
-            k=2)
-        seqs_to_consider = {0: set(range(len(seqs)))}
-        self.gen_gs.mismatches = 1
-
-        # The best guide at start=2 is 'TT', but if we require
-        # 'C' to flank on the 5' end, the best is 'AA'
-        self.gen_gs.required_flanking_seqs = ('C', None)
-        self.assertEqual(self.gen_gs.construct_oligo(2, guide_length,
-                                                     seqs_to_consider),
-                         ('AA', {0,1,9,10}, 4))
-
-        # The best guide at start=2 is 'TT', but if we require
-        # 'C' to flank on the 5' end, the best is 'AA'
-        # Now if we require 'M' on the 5' end, 'TT' will be the best guide
-        self.gen_gs.required_flanking_seqs = ('M', None)
-        self.assertEqual(self.gen_gs.construct_oligo(2, guide_length,
-                                                     seqs_to_consider),
-                         ('TT', {2,3,4,5,6,7,8}, 7))
-
     def tearDown(self):
         # Re-enable logging
         logging.disable(logging.NOTSET)
